make location --build-dir return build directory
Currently when you use `dev-build` or `develop` with cmake packages you
have no clue where the build folder really is if you just have the name
of the spec but not the hash. This is a bit annoying when you dev-build
two versions of the same sources and end up with two folders
`sources/spack-build-abcdef` and `sources/spack-build-012345`. This is a
result of https://github.com/spack/spack/pull/18574.

This PR makes `spack location --build-dir`return the sensible folder for
cmake packages and others that support out of source builds. This way
it's easier to programmatically `cd` into the actual build folder and
run make or ctest by hand, without having to figure out the hash.